+++ CHANGELOG.md
@@ -1,14 +1,16 @@
 # Last changes:
 
 
-Version `v1.4.1 (25 Jun'25)`
+Version `v1.4.2b2 (2 Jul'25)`
 
   - BUGFIX: Preferences: "Test" button's issues
   - BUGFIX: Preferences: modified EG field is cleared at OK (without restart)
   - BUGFIX: Statistics: no cnt0 counted at val<0
   - NEW   : argument *-debug*  for logs to ~/TMP (+ disables 1 app instance check)
   - NEW   : "Home" button (Ctrl+H) redraws also the diagram to see current week
-  - CHANGE: text cell's value: leading +/- sets value (not incrs/decrs)
+  - NEW   : text cell's value in the cell's tip
+  - NEW   : text cell's value: leading +2, +3.. -2, -3.. sets value
+  - CHANGE: text cell's value: leading +/- sets value (not incr/decr); default is 0
   - CHANGE: Round proc: rounding negative numbers by abs
   - CHANGE: Statistics: at opening, current date range & table
   - CHANGE: Statistics: text items' & EG's values formatted as 999